fix_norm_vector_ope
