+++ srcs/draw/test_intersection/rt_cy_intersec.c
@@ -24,7 +24,7 @@ static double	rt_calc_abd(t_cylinder *cy, t_ray ray, double *a, double *b)
 	t_3d_vec	s_cy_c;
 	double		c;
 
-	d = rt_vec_copy(ray.unit_d_vec);
+	d = ray.unit_d_vec;
 	s_cy_c = rt_vec_sub(ray.start, cy->center_p_vec);
 	*a = rt_vec_dot(d, d) - pow(rt_vec_dot(d, cy->unit_orient_vec), 2.0);
 	*b = 2.0 * (rt_vec_dot(d, s_cy_c) - (rt_vec_dot(d, cy->unit_orient_vec))
